This is my commit message

+++ bundles/org.openhab.binding.modbus.lambda/src/main/java/org/openhab/binding/modbus/lambda/internal/LambdaConfiguration.java
@@ -26,7 +26,7 @@ public class LambdaConfiguration {
     /**
      * Refresh interval in seconds
      */
-    private long refresh = 30;
+    private int refresh = 30;
 
     private int maxTries = 3;// backwards compatibility and tests
